@@ -36,10 +36,13 @@ e.g. If you have a core plugin that has a 'hello' command, installing a user-ins
     for (let name of argv) {
       if (aliases[name] === null) this.error(`${name} is blacklisted`)
       name = aliases[name] || name
-      let p = parsePlugin(name)
+      let p = await this.parsePlugin(name)
       let plugin
       if (p.type === 'npm') {
         cli.action.start(`Installing plugin ${chalk.cyan(this.plugins.friendlyName(p.name))}`)
+        await this.config.runHook('plugins:preinstall', {
+          plugin: p
+        })
         plugin = await this.plugins.install(p.name, {tag: p.tag, force: flags.force})
       } else {
         cli.action.start(`Installing plugin ${chalk.cyan(p.url)}`)
@@ -48,18 +51,19 @@ e.g. If you have a core plugin that has a 'hello' command, installing a user-ins
       cli.action.stop(`installed v${plugin.version}`)
     }
   }
-}
 
-function parsePlugin(input: string): {name: string, tag: string, type: 'npm'} | {url: string, type: 'repo'} {
-  if (input.includes('@') && input.includes('/')) {
-    input = input.slice(1)
-    let [name, tag = 'latest'] = input.split('@')
-    return {name: '@' + name, tag, type: 'npm'}
-  } else if (input.includes('/')) {
-    if (input.includes(':')) return {url: input, type: 'repo'}
-    else return {url: `https://github.com/${input}`, type: 'repo'}
-  } else {
-    let [name, tag = 'latest'] = input.split('@')
-    return {name, tag, type: 'npm'}
+  async parsePlugin(input: string): Promise<{name: string, tag: string, type: 'npm'} | {url: string, type: 'repo'}> {
+    if (input.includes('@') && input.includes('/')) {
+      input = input.slice(1)
+      let [name, tag = 'latest'] = input.split('@')
+      return {name: '@' + name, tag, type: 'npm'}
+    } else if (input.includes('/')) {
+      if (input.includes(':')) return {url: input, type: 'repo'}
+      else return {url: `https://github.com/${input}`, type: 'repo'}
+    } else {
+      let [name, tag = 'latest'] = input.split('@')
+      name = await this.plugins.maybeUnfriendlyName(name)
+      return {name, tag, type: 'npm'}
+    }
   }
 }
