@@ -2,46 +2,46 @@ import {expect, test} from '../../test'
 
 describe('command', () => {
   test
-  .command(['plugins:install', '@oclif/example-plugin-ts'], {reset: true})
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('@oclif/example-plugin-ts '))
-  .stdout()
-  .command(['hello'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('hello world'))
-  .command(['plugins:uninstall', '@heroku-cli/plugin-@oclif/example-plugin-ts'])
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
-  .it('installs and uninstalls @oclif/example-plugin-ts')
+    .command(['plugins:install', '@oclif/example-plugin-ts'], {reset: true})
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('@oclif/example-plugin-ts '))
+    .stdout()
+    .command(['hello'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('hello world'))
+    .command(['plugins:uninstall', '@heroku-cli/plugin-@oclif/example-plugin-ts'])
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
+    .it('installs and uninstalls @oclif/example-plugin-ts')
 
   test
-  .command(['plugins:install', 'aliasme'], {reset: true})
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('oclif-debug'))
-  .stdout()
-  .command(['debug'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('debug'))
-  .command(['plugins:uninstall', 'oclif-debug'])
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
-  .it('installs via an alias')
+    .command(['plugins:install', 'aliasme'], {reset: true})
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('oclif-debug'))
+    .stdout()
+    .command(['debug'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('debug'))
+    .command(['plugins:uninstall', 'oclif-debug'])
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
+    .it('installs via an alias')
 
   test
-  .command(['plugins:install', 'jdxcode/oclif-debug'], {reset: true})
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('oclif-debug'))
-  .stdout()
-  .command(['debug'], {reset: true})
-  .do(output => expect(output.stdout).to.contain('debug'))
-  .command(['plugins:uninstall', 'oclif-debug'])
-  .stdout()
-  .command(['plugins'], {reset: true})
-  .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
-  .it('installs and uninstalls jdxcode/oclif-debug')
+    .command(['plugins:install', 'jdxcode/oclif-debug'], {reset: true})
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('oclif-debug'))
+    .stdout()
+    .command(['debug'], {reset: true})
+    .do(output => expect(output.stdout).to.contain('debug'))
+    .command(['plugins:uninstall', 'oclif-debug'])
+    .stdout()
+    .command(['plugins'], {reset: true})
+    .do(output => expect(output.stdout).to.equal('no plugins installed\n'))
+    .it('installs and uninstalls jdxcode/oclif-debug')
   // test
   // .command(['plugins:install', 'heroku-debug@beta'], {reset: true})
   // .stdout()
