@@ -2,7 +2,7 @@ import {test} from '../../test'
 
 describe('command', () => {
   test
-  .command(['plugins:uninstall', 'foobar'], {reset: true})
-  .catch(/foobar is not installed/)
-  .it('uninstall non-existent plugin')
+    .command(['plugins:uninstall', 'foobar'], {reset: true})
+    .catch(/foobar is not installed/)
+    .it('uninstall non-existent plugin')
 })
